chore(sync): Fase 4 — robustness cleanup (ordering tie-break, folio BigInt, DebugLog pruning, stuck-row diagnostic)
- sp_GetPendingChanges: ORDER BY Timestamp ASC, Id ASC. Without the Id tie-break,
  rows sharing the same millisecond Timestamp could reorder between polls and an
  early row might never reach the TOP(@MaxResults) window (starvation). Id is the
  IDENTITY PK, so this makes the ordering stable and insertion-monotonic.
- producer: bind folio as sql.BigInt (was sql.Int) in DB lookups — tempcheques.folio
  is BIGINT, so high-volume venues risked silent Int overflow. (Adapter command-path
  bindings tracked as a deferred LOW item.)
- sp_CleanupOldTrackingRecords: also prune AvoqadoDebugLog (Timestamp < cutoff);
  sp_ApplyPartialPayment writes ~10 rows/call and nothing pruned it before.
- 03-DIAGNOSTICS: new "STUCK / DROPPED TRACKING ROWS" alert — lists rows with
  ProcessedAt IS NULL AND RetryCount >= 5 (silently excluded by sp_GetPendingChanges'
  RetryCount < 5 filter → never delivered, never retried). Turns a silent event loss
  into an alert. Plus an AvoqadoDebugLog growth watch.
- Docs: ✅ notes on the four cleanup gaps + a "Diferido (Fase 4+)" list (adapter folio,
  dead code, cash H-11, trigger-guard optimization, stop-shrinking-total redesign).

All SQL deployed to the avo test DB and confirmed to run. tsc/prettier clean.

+++ src/components/producer.ts
@@ -509,12 +509,12 @@ async function processOrderChangeV10(change: ChangeNotification, venueId: string
     if (idturno && idturno !== 'null' && isValidIdTurno) {
       log.info(`[Order Processor] Buscando orden con idturno ${numericIdTurno} y folio ${folio}`)
       request.input('idturno', sql.BigInt, numericIdTurno)
-      request.input('folio', sql.Int, folio)
+      request.input('folio', sql.BigInt, folio)
       query = 'SELECT * FROM tempcheques WHERE idturno = @idturno AND folio = @folio'
     } else {
       // Si idturno es nulo o una cadena vacía, buscamos la orden sin turno asignado.
       log.info(`[Order Processor] Buscando orden con idturno NULO y folio ${folio}`)
-      request.input('folio', sql.Int, folio)
+      request.input('folio', sql.BigInt, folio)
       query = 'SELECT * FROM tempcheques WHERE idturno IS NULL AND folio = @folio'
     }
 
@@ -555,7 +555,7 @@ async function processOrderChangeV10(change: ChangeNotification, venueId: string
     {
       const paymentsRes = await pool
         .request()
-        .input('folio', sql.Int, folio)
+        .input('folio', sql.BigInt, folio)
         .query('SELECT idformadepago, importe, propina, referencia FROM tempchequespagos WHERE folio = @folio')
 
       if (paymentsRes.recordset.length > 0) {
@@ -673,7 +673,7 @@ async function processOrderChangeV11(
     {
       const paymentsRes = await pool
         .request()
-        .input('folio', sql.Int, posData.folio)
+        .input('folio', sql.BigInt, posData.folio)
         .query('SELECT idformadepago, importe, propina, referencia FROM tempchequespagos WHERE folio = @folio')
 
       if (paymentsRes.recordset.length > 0) {
